+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/media/MediaTooLargeException.kt
@@ -1,0 +1,3 @@
+package net.folivo.trixnity.client.media
+
+object MediaTooLargeException : IllegalStateException("your file is larger then the servers max upload size")