+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/room/OutboxMessageEventHandler.kt
@@ -17,6 +17,7 @@ import net.folivo.trixnity.client.CurrentSyncState
 import net.folivo.trixnity.client.MatrixClientConfiguration
 import net.folivo.trixnity.client.flattenNotNull
 import net.folivo.trixnity.client.media.MediaService
+import net.folivo.trixnity.client.media.MediaTooLargeException
 import net.folivo.trixnity.client.room.outbox.OutboxMessageMediaUploaderMappings
 import net.folivo.trixnity.client.room.outbox.findUploaderOrFallback
 import net.folivo.trixnity.client.store.RoomOutboxMessage
@@ -128,17 +129,26 @@ class OutboxMessageEventHandler(
                     outboxMessage.keepMediaInCache,
                 ).getOrThrow()
             }
-        } catch (exception: MatrixServerException) {
-            val sendError = when (exception.statusCode) {
-                HttpStatusCode.Forbidden -> SendError.NoMediaPermission
-                HttpStatusCode.PayloadTooLarge -> SendError.MediaTooLarge
-                HttpStatusCode.BadRequest -> SendError.BadRequest(exception.errorResponse)
-                HttpStatusCode.TooManyRequests -> throw exception
+        } catch (exception: Exception) {
+            val sendError = when (exception) {
+                is MatrixServerException -> when (exception.statusCode) {
+                    HttpStatusCode.Forbidden -> SendError.NoMediaPermission
+                    HttpStatusCode.PayloadTooLarge -> SendError.MediaTooLarge
+                    HttpStatusCode.BadRequest -> SendError.BadRequest(exception.errorResponse)
+                    HttpStatusCode.TooManyRequests -> throw exception
+                    else -> {
+                        log.error(exception) { "could not upload media" }
+                        SendError.Unknown(exception.errorResponse)
+                    }
+                }
+
+                MediaTooLargeException -> SendError.MediaTooLarge
                 else -> {
                     log.error(exception) { "could not upload media" }
-                    SendError.Unknown(exception.errorResponse)
+                    SendError.Unknown(null, exception.message ?: exception::class.simpleName ?: "unknown")
                 }
             }
+
             roomOutboxMessageStore.update(outboxMessage.transactionId) {
                 it?.copy(sendError = sendError)
             }