+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/ServerVersions.kt
@@ -1,9 +1,0 @@
-package net.folivo.trixnity.client.store
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ServerVersions(
-    val versions: List<String>,
-    val unstableFeatures: Map<String, Boolean>,
-)