+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/InMemoryRepository.kt
@@ -60,7 +60,7 @@ abstract class InMemoryMapRepository<K1, K2, V> : MapRepository<K1, K2, V> {
 }
 
 class InMemoryAccountRepository : AccountRepository, InMemoryMinimalRepository<Long, Account>()
-class InMemoryServerVersionsRepository : ServerVersionsRepository, InMemoryMinimalRepository<Long, ServerVersions>()
+class InMemoryServerDataRepository : ServerDataRepository, InMemoryMinimalRepository<Long, ServerData>()
 class InMemoryOutdatedKeysRepository : OutdatedKeysRepository, InMemoryMinimalRepository<Long, Set<UserId>>()
 class InMemoryDeviceKeysRepository : DeviceKeysRepository,
     InMemoryMinimalRepository<UserId, Map<String, StoredDeviceKeys>>()