+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/createInMemoryRepositoriesModule.kt
@@ -7,7 +7,7 @@ fun createInMemoryRepositoriesModule() = module {
     single<RepositoryTransactionManager> { NoOpRepositoryTransactionManager }
 
     singleOf<AccountRepository>(::InMemoryAccountRepository)
-    singleOf<ServerVersionsRepository>(::InMemoryServerVersionsRepository)
+    singleOf<ServerDataRepository>(::InMemoryServerDataRepository)
     singleOf<OutdatedKeysRepository>(::InMemoryOutdatedKeysRepository)
     singleOf<DeviceKeysRepository>(::InMemoryDeviceKeysRepository)
     singleOf<CrossSigningKeysRepository>(::InMemoryCrossSigningKeysRepository)