+++ trixnity-client/trixnity-client-repository-exposed/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/exposed/createExposedRepositoriesModule.kt
@@ -20,7 +20,7 @@ suspend fun createExposedRepositoriesModule(
     newSuspendedTransaction(Dispatchers.IO, database) {
         val allTables = arrayOf(
             ExposedAccount,
-            ExposedServerVersions,
+            ExposedServerData,
             ExposedCrossSigningKeys,
             ExposedDeviceKeys,
             ExposedGlobalAccountData,
@@ -53,7 +53,7 @@ suspend fun createExposedRepositoriesModule(
         single { database }
         singleOf(::ExposedRepositoryTransactionManager) { bind<RepositoryTransactionManager>() }
         singleOf(::ExposedAccountRepository) { bind<AccountRepository>() }
-        singleOf(::ExposedServerVersionsRepository) { bind<ServerVersionsRepository>() }
+        singleOf(::ExposedServerDataRepository) { bind<ServerDataRepository>() }
         singleOf(::ExposedOutdatedKeysRepository) { bind<OutdatedKeysRepository>() }
         singleOf(::ExposedDeviceKeysRepository) { bind<DeviceKeysRepository>() }
         singleOf(::ExposedCrossSigningKeysRepository) { bind<CrossSigningKeysRepository>() }