+++ trixnity-client/trixnity-client-repository-indexeddb/src/jsTest/kotlin/net/folivo/trixnity/client/store/repository/indexeddb/RepositoryTestSuite.kt
@@ -22,7 +22,7 @@ class RepositoryTestSuite : ShouldSpec({
                 IndexedDBRepositoryTransactionManager(get(), allStoreNames, testMode = true)
             }
             singleOf(::IndexedDBAccountRepository) { bind<AccountRepository>() }
-            singleOf(::IndexedServerVersionsRepository) { bind<ServerVersionsRepository>() }
+            singleOf(::IndexedServerDataRepository) { bind<ServerDataRepository>() }
             singleOf(::IndexedDBCrossSigningKeysRepository) { bind<CrossSigningKeysRepository>() }
             singleOf(::IndexedDBDeviceKeysRepository) { bind<DeviceKeysRepository>() }
             singleOf(::IndexedDBGlobalAccountDataRepository) { bind<GlobalAccountDataRepository>() }